spec: the recurrent refold is private to the module that wraps it
