+++ CLAUDE.md
@@ -333,7 +333,7 @@ hand — keeps the CI gate and the local gate identical.
 | `make check-spec-sampling` | CI gate (in `make ci`): every speculative generation driver — a `pub fn` taking `step_fn: &mut dyn FnMut(&ProbeStep)` — is handed the request's sampler and reads it, and every drafter arm of the server's dispatch passes it. A loop that takes no sampler decodes greedily, does not warn, and returns fluent text, so the caller's temperature is lost with nothing anywhere saying so. This gate cannot tell a right distribution from a wrong one; `crates/rmlx-models/tests/spec_sampled_distribution.rs` does that, on one pair under `make gpu-test`. A scan matching no driver is exit 2, not a pass. |
 | `make check-spec-sampling-fixtures` | CI gate (in `make ci`): recall test for the above, 10 synthetic scan roots, each asserting the reason as well as exit 1 vs exit 2. |
 | `make check-spec-charge` | CI gate (in `make ci`): each speculative round loop names one phase-charge decision and names it everywhere the decision is read — the `charge` argument of every `rollback_round` call inside the loop and the `charged:` field of every `RoundPhases` event and of the `RoundTotals` it hands the one recorder are the same token, and no fn that is not a round loop writes the field — fn bodies only, so a decision parked in a module-level constant is outside it; the recorder carries the value across as a destructured `charged,`, and `crates/rmlx-models/src/speculative/round_common_tests.rs` is what reads that it arrives — and the census over the seven loops is exactly `charge_phases` three times and `false` four. A loop that charges its phases and records that it did not re-attributes its own work to the drafter with every token and every count identical, and no other observable in the tree can see it: the equivalence pairs read the answer, the accept counters read the aggregate, and the per-round stream reads `charged` as `false` on both sides by construction, because a capture that enabled the switch would be measuring a different, slower run. Fewer loops than the tree ships, or a call whose arguments the scan cannot read back, is exit 2. |
-| `make check-spec-charge-fixtures` | CI gate (in `make ci`): recall test for the above, 29 synthetic scan roots carrying the tree's own census, each asserting the reason as well as exit 1 vs exit 2 — including the two shapes that drop a loop out of the derived population, which must read as a lost loop and not as a census that moved, and the shared seed emit, which carries a driver's signature and is kept out by naming no totals of its own. |
+| `make check-spec-charge-fixtures` | CI gate (in `make ci`): recall test for the above, 30 synthetic scan roots carrying the tree's own census, each asserting the reason as well as exit 1 vs exit 2 — including the two shapes that drop a loop out of the derived population, which must read as a lost loop and not as a census that moved, and the shared seed emit, which carries a driver's signature and is kept out by naming no totals of its own. |
 | `make check-doc-source-citations` | CI gate (in `make ci`): every `crates/...` path cited in `docs/` resolves. Covers paths only, not identifiers — telling a test name from a variable needs a parser. |
 | `make debt-report` | Advisory (non-failing): sibling-file/fn similarity ("twins") in `crates/rmlx-kv-quant` and `crates/rmlx-models` — printing the speculative round-loop drivers as a named group regardless of threshold, since that pairing is the standing violation the twin rule names. The group is *discovered*, not a literal name list: a driver is any fn under `crates/rmlx-models/src/speculative/` whose signature carries `step_fn: &mut dyn FnMut(&ProbeStep)`, the same rule `scripts/check_spec_sampling.sh` uses — generalised to any visibility, which is what surfaces the private cached loops a `pub fn`-only list misses. Four debt counters (`#[allow(` sites, debt-marker comments, `check-*` Make targets, oversized files without `LOC-exempt`, each stating its own file population), the churn (summed over commits, not a net diff) since the last tag, and `docs/*.md` files over 200 KB. Also runs at the end of `make ci` and in hosted CI (advisory, non-blocking). |
 | `make debt-report-selftest` | CI gate (in `make ci` and hosted CI): the above over synthetic fixtures — a planted twin pair is reported and a same-naming-shape pair just under the threshold is not, the round-loop group is discovered by signature (not by name — a renamed driver is found under its new name, with the total count unaffected), two named `fn …: NN.N% shared` lines are asserted from the twin pair, all four debt counters are asserted against planted fixtures (one `#[allow(`, one debt comment in source and one in a `_tests.rs` file, a Makefile with two `check-*:` targets, an oversized file and its `LOC-exempt`-marked twin), and a synthetic two-commit, one-tag git repo (plus a top-level `README.md` alongside a `docs/*.md` file) proves the churn figure and its "*.md" pathspec are both exercised. Asserts what each case found, not just that the tool ran. |
